requirements.txt
numpy==1.18.5
plotly==4.14.1
streamlit==0.73.1
pandas==1.0.5
Pillow==8.1.0
spacy==2.1.8
nltk
scikit-learn>=0.24
Pylons==1.0
wordcloud
